Applying the Unlicense to the repository
After a bit of reading on https://choosealicense.com/licenses/unlicense/
I felt this was the most applicable license for this repository. This
repository is mostly intended as a public journal on how I managed my
personal IT infrastructure. I hope people look at it and chat with me
about how it works. This should at least limit my liability if someone
tries to use it and negatively impacts them and they get angry with me.

This resolves #18