Add reset_prompt() function
This function will bring the PROMPT variable back to a known baseline.
The reason for this is to reset any PROMPT mangaling that other shell
extensions perform.

For example, I might have enabled a k8s_PS1. Then I work on things
unrelated to k8s workflows so I can run reset_prompt to remove the
PROMPT noise for things that aren't relevant to what I am working on.