Bring back ssh-copy-id to checklist #35
Discovered with @konradfischer during the latest repave. Ansible doesn't
allow for specifying passwords for password based auth in an ssh
session.

After further discussion with @mrdavidlaing, noted that the `ansible.sh`
was unnecessarily passing `--ask-become-pass`. My previous understanding
was this this asked the password for to establish an ssh session, but
it's the password needed for when `sudo` prompts for a password.

These will be cleaned up in a subsequent commit(s).