Automatically ensure ssh-agent is running
Not sure if the automation is really necessary here. Launching the agent
when I needed it works, but it did result in excess ssh-agent processes
running over time.

I'm tempted revert back to starting the agent only when I need it, but
wrap that in a function that will load the env if an agent already
happens to be running to spare the process litter.