Fix obtaining ssh key passphrase fetching in linux
The DISPLAY env variable trick wasn't working in Ubuntu. I was happy to
learn that in 2020 a new env var was introduced that can force
SSH_ASKPASS to be used.

Since I haven't tested this on my macOS machine, I left the DISPLAY
variable there because I'm pretty sure macOS is using a different
openssh package.