Extend DHCP leases for home network
When observing logs from the PI, there's a lot of kea activity. Given
that my home network doesn't change much, I felt that I didn't need to
renew leases so often.

This article gave me some insight on the change:
https://lazyadmin.nl/home-network/dhcp-lease-time/

I'm not sure if the renew and rebind timers make sense. I adjusted them
to the same relative value of the valid-lifetime.