Add alacritty-colorscheme
This was surprisingly easy to setup and now creates consisten color
schemes across alacritty, nvim, and delta!

This leaves ~/.vimrc_background to inform nvim what colorscheme to load.
This may result in some awkwardness on a fresh system.

`alacritty-colorscheme status` will display which theme is being used.

Since `alacritty-colorscheme` modifies the alacritty.yml file with the
chosen theme, this may result in some git diff noise. I'll look to see
if I can set this up as an include so that this doesn't result in a
change to be staged. One idea is to always switch to a specific theme
when running `coalesce_this_machine`.

Had to write `delta-themed` to auto-populate the --syntax-theme paramter
based on the current theme.

Now to switch to Alacritty on Windows as well