Switch from kea to dnsmasq for dhcpd
Thankfully this was pretty straightforward. The dnsmasq.conf is much
smaller than the kea config and is easy to understand.

I think I'll use /etc/dnsmasq.d in the future to avoid messing directly
with /etc.

So far this configuration is working with my Macbook Air, Windows PC,
and Android phone.

Given that dnsmasq is equivalent in simplicity from a configuration
management perspective, I'm going to look at migrating to it for DNS and
remove `unbound`.

This fixes #42