Resolve remaining ansible-lint errors
I like the alternative to the sudoer update with this. I find *.d/ type
configuration more intuitive and easier for automation.

Removed the test at the end since it is a bit superflous. If the
automation worked until the end, we should be good to go. If not, I
would find out soon when running the main playbook doesn't work.