Ensure main playbook comes first
Not sure I like this subshell execution. Hard coding the file names
feels like it would be a less magical, and more understandable. I'm not
saving time or effort through the execution of `find`. What it does is
enable renaming the playbook files without having to modify the options
to `gum choose`. By adding the `sort -r`, this introduces an implicity
coupling based of an assumed file name strategy.