Extract certbot actions outside of nginx
The nginx configuration is a bit complicated and I am exploring how to
simplify the design.

This commit will break a new repave because of the ordering of the
tasks, but I'll look at fixing that next. This is because nginx will
attempt to be stopped before nginx is installed.