Remove resolv.conf changes for dnsresolver
I believe the first time I introduced this, it was being too defensive.
When the PI first boots, it'll have a dynamically assigned IP and will
get DNS resolver settings from that.

If this still is a problem in the future, it'll be resolved (pun
intended) when I have 2 resolvers in my LAN.