Move the remaining roles into tasks
In the end, I think think increases the simplicity of the playbook,
makes it easier to follow, and I still got to learn how to make a role.

My takeway is that it was nice to abstract away the Handler to a role,
but didn't get any other benefits over a separate task file.

Since this playbook isn't too general purpose, creating roles with vars
and defaults doesn't make too much sense. Hardcoding for simplicity is
still good until abstraction is warranted.