Fix ssh-agent startup and key loading
The .zshrc has a function to ensure the ssh-agent is started whenever I
launch into my WSL shell. The functions fail if the .ssh directory
doesn't exist, and I was previously reconciling that issue manually.

Now that the new WSL instance is running Ubuntu Jammy, the bc command
isn't installed by default so adding that to the OS package installation
list.

Relates to #48

Signed-off-by: Scott Muc <scott@scottmuc.com>