Remove go related installation
This used to be required for the terraform gandi provider, but that's a
proper release now, so I don't need to build it.

I'm also not doing any go dev or using any `go install` binaries. I can
use `.tool-versions` in the appropriate repositories if I need to use
golang again.

Related: #49