Simplify alacritty config and setup (wsl)
Rather than having a commented out config file, I'm leaving a reference
to valid configuration. I'm not changing much from the defaults, and
this highlights that.

I've chosen to not symlink the config, because the file changes whenever
a color scheme is selected. I don't adjust this configuration opten and
having the a changed file in my repo was getting frustrating.

This does mean everytime I run `coalesce_this_machine` the theme will be
over-written... which I'm ok with for now.

The windows alacritty setup was responsible for setting up the config
directory in windows, but now moved that to the theme switcher.

Issue: #50