Navidrome configuration cleanup
A functional change here is that I've disabled the scheduled scanning of
my library. It doesn't change often so I'm happy to trigger a rescan
manually whenever I add new files.

Configuration in the systemd service definition was moved to the
configuration file. I'd like to keep that file as application context
free as possible. Just know how to start the process and tell it where
the configuration file lives.

Updated github release URL to use the navidrome org repo rather than
following the deluan redirected URL.

Related to #59