Improve repaving steps
By setting a static IP, I can ignore the disable DHCPD step. Perhaps
this might need to be the first step, but I'll go with this for now.

Also add a step to redeploy logging after a reboot with details on how
to manage this if the step persists for too long.

Issue: #61