Switch to Lazy from Packer for nvim packages
Thanks @thenbe for the pointers!

Decided to not check in the `lazy-lock.json` for now. Happy to keep
upgrading to the latest versions. That being said, maintaining this file
would leave a potentially useful version paper trail. I'll add it if it
turns out it is valuable.

Needed to require the plugins code after mapping my leader key. Learnt
about this with this kind comment in kickstart.nvim:
https://github.com/nvim-lua/kickstart.nvim/blob/master/init.lua#L42

Issue: #63