Update static ip assignment for bookworm
bookworm no longer has a dhcpcd service. It looks like it altered it's
dhcp stack to use the isc tools. This change allowed the playbook to
progress, but will need to investigate further if this is sufficient for
disabling a dhcp client going forward (and for how to set a static ip).

Issue: #65