Begin diverging from kickstart.nvim
I'm growing more comfortable with the Lua style configuration syntax and
directory structure. The goal is to start pruning Kickstart and
configuring it for what I need.

The configuration has been copied to lua/kickstart/init.lua as a way of
being able to do a require("kickstart.init") if I ever want to get back
to that configuration.

With this, I think much of the commenting can be stripped and I can
point to the original Kickstart config for the helpful comments.