Enable git repos to persist across repaves
Changes the git user to have an ephemeral home directory on the SD card
which will get wiped during repaving. Repositories stay on the external
USB drive and get symlinked in the git home.

This change also augments the path to the repositories to remove the
redundant scottmuc/ path as well as renames the scottmuc.com.git to
website.git. The latter change might still benefit from being a bit more
explicit in what domain the website is for.

Cloning now looks like:

git clone git@pi:website.git

I always fixed the idempotent check for the initalizing of the bare
repositories. I tested this by deleting the user (userdel -r git) and
rerunning the playbook. The repositories were already initialized and
the task was skipped, but the symlinks did get recreated.

Issue: #71