Resolve git.scottmuc.com to private IP at home
This creates an explicit dependency on unbound for this task. Given
this is all orchestrated in a single playbook, this is fine to me.

This configuration is more related to configuring the githost than it
is general purpose DNS, so I felt adding it to this task is clearer.

Issue: #71