Bind unbind to all interface
This is relevant for repaving because installation will occur on
a random IP. This might indicate moving baseline network configuration
to the bootstrap playbook, or create a new playbook for stable network
bootstrapping.

Also, remove avahi since it's unneccessary and it conflicts on port 5353
for UDP which can mess up local hostname resolving.

Issue: #79