Updated node bootstrap for when sudo isn't available
The machine is currently named gimli, but that'll change soon. This was
an interested chicken and the egg problem. The Raspian image already has
sudo installed, and the user it creates as part of image creation can
sudo correctly.

This reminds me that I need to update root password on the XPS
immediately. I don't even know what the root password is on the PI.