Fetch pocket saves in batches of 30
The API has changed in a way that it requires looping through a batch of
requests to get the full dataset. Thankfully jq -s makes merging all the
responses into 1 file was easy!

deep-pockets stats command gives me an article count that matches what
the total field in the response says.

Given that pocket is shutting down, this is probably the last iteration
of the script.