Widen the allowed host for music export
When Pippin is repaved, the pippin name doesn't resolve correctly until
it's fully configured and reboot. This prevents Pippin from mounting the
nfs export... so this change makes it mountable regardless of the IP
Pippin is using.

The sync option was removed since it doesn't do anything useful for a
read-only export.

Issue: #92