Verify the song played for 5 seconds
The test config now verifies all env vars are set with a custom
message for the user in case a var is missing. I couldn't use
playwrights built-in custom message because cucumber seems to
swallow it (several threads on github and stackoverflow found
with people requesting it). But using a try catch works.

The tests now verify that the song played at least 5 secs by
checking timestamps.

Issue: #91