Setup nfs on erebor
This was as easy as following the Handbook docs:
https://docs.freebsd.org/en/books/handbook/network-servers/#network-nfs

The gotcha was that ansible understands that the service is nfsd but
what needs to go into /etc/rc.conf to enable it is nfs_server.

I'm not 100% sure about the exports. I think I'll switch this to be
very host specific. I think I only want the following machines being
able to access the NFS:

* merry (unix read read only access for access to files)
* gandalf (windows/wsl read write access for backups)
* frodo (unix read write access for backups)

Issue: #100