Restart mountd and other nfs processees when exports change
I was having issues with mounting from frodo after I rebooted erebor.
This clued me to the fact that nfsd isn't the only thing that needs
to restart if /etc/exports change.

Since frodo is not a user on erebor, I needed to use -mapall=nfsrw:nfsrw

Deleted the readonly share since it's not needed now. I'll revisit that
when I setup other servers that need only read access to the data.

Issue: #100